node-main/deps/v8/test/inspector/debugger/wasm-instrumentation-breakpoint-expected.txt

Test instrumentation breakpoints in wasm.

Running test: testBreakInStartFunction
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Compiling wasm module.
Paused at v8://test/compile_module with reason "instrumentation".
Hit breakpoints: []
Instantiating module.
Paused at v8://test/instantiate with reason "instrumentation".
Hit breakpoints: []
Paused at wasm://wasm/341d515e with reason "instrumentation".
Script wasm://wasm/341d515e byte offset 26: Wasm opcode 0x01 (kExprNop)
Hit breakpoints: []
Instantiating a second time (should trigger no breakpoint).
Paused at v8://test/instantiate2 with reason "instrumentation".
Hit breakpoints: []
Done.

Running test: testBreakInStartFunctionCompileTwice
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Instantiating module.
Paused at v8://test/instantiate with reason "instrumentation".
Hit breakpoints: []
Paused at wasm://wasm/341d515e with reason "instrumentation".
Script wasm://wasm/341d515e byte offset 26: Wasm opcode 0x01 (kExprNop)
Hit breakpoints: []
Instantiating a second time (should trigger another breakpoint).
Paused at v8://test/instantiate with reason "instrumentation".
Hit breakpoints: []
Paused at wasm://wasm/341d515e with reason "instrumentation".
Script wasm://wasm/341d515e byte offset 26: Wasm opcode 0x01 (kExprNop)
Hit breakpoints: []
Done.

Running test: testBreakInExportedFunction
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Instantiating wasm module.
Paused at v8://test/instantiate with reason "instrumentation".
Hit breakpoints: []
Calling exported function 'func' (should trigger a breakpoint).
Paused at v8://test/call_func with reason "instrumentation".
Hit breakpoints: []
Paused at wasm://wasm/aa8b4c12 with reason "instrumentation".
Script wasm://wasm/aa8b4c12 byte offset 33: Wasm opcode 0x01 (kExprNop)
Hit breakpoints: []
Calling exported function 'func' a second time (should trigger no breakpoint).
Paused at v8://test/call_func with reason "instrumentation".
Hit breakpoints: []
Done.

Running test: testBreakOnlyWithSourceMap
Setting instrumentation breakpoint for source maps only
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Instantiating wasm module without source map.
Calling exported function 'func' (should trigger no breakpoint).
Instantiating wasm module with source map.
Calling exported function 'func' (should trigger a breakpoint).
Paused at wasm://wasm/b2871d1e with reason "instrumentation".
Script wasm://wasm/b2871d1e byte offset 33: Wasm opcode 0x01 (kExprNop)
Hit breakpoints: []
Done.

Running test: testRemoveBeforeCompile
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Remove instrumentation breakpoint..
Compiling wasm module.
Instantiating module should not trigger a break.
Done.

Running test: testRemoveBeforeInstantiate
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Compiling wasm module.
Paused at v8://test/compile_module with reason "instrumentation".
Hit breakpoints: []
Remove instrumentation breakpoint..
Instantiating module should not trigger a break.
Done.

Running test: testRemoveAfterOnePause
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Compiling wasm module.
Paused at v8://test/compile_module with reason "instrumentation".
Hit breakpoints: []
Instantiating module should trigger a break.
Paused at v8://test/instantiate with reason "instrumentation".
Hit breakpoints: []
Paused at wasm://wasm/341d515e with reason "instrumentation".
Script wasm://wasm/341d515e byte offset 26: Wasm opcode 0x01 (kExprNop)
Hit breakpoints: []
Remove instrumentation breakpoint..
Compiling another wasm module.
Instantiating module should not trigger a break.
Done.

Running test: testDisableEnable
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Compiling wasm module.
Paused at v8://test/compile_module with reason "instrumentation".
Hit breakpoints: []
Disable debugger..
Enable debugger
Instantiating module should not trigger a break.
Done.
